Revert "RavenDB-7070 - move tests to FastTests"
This reverts commit ee47145b0b8f2dca4c01b34a591193fdd1a1160c.
